+++ website/components/PropTable.tsx
@@ -1,28 +1,36 @@
 import React from 'react';
 import { css } from 'emotion';
+import Card from '@leafygreen-ui/card';
 import { Table, TableHeader, Row, Cell } from '@leafygreen-ui/table';
-import { Subtitle, InlineCode } from '@leafygreen-ui/typography/dist';
+import { Subtitle, InlineCode } from '@leafygreen-ui/typography';
 import { OneOf } from '@leafygreen-ui/lib';
+import { useViewportSize } from '@leafygreen-ui/hooks';
+import { breakpoints } from '@leafygreen-ui/tokens';
 import PropDefinition from 'components/PropDefinition';
 import TypographyPropTable from 'components/TypographyPropTable';
 import formatType from 'utils/formatType';
 import { mq } from 'utils/mediaQuery';
+import { pageContainerWidth } from 'styles/constants';
 
 const tableWrapper = css`
   ${mq({
     marginLeft: ['-24px', 'unset'],
     marginRight: ['-24px', 'unset'],
     overflow: ['hidden', 'unset'],
+    width: [
+      'inherit',
+      'inherit',
+      'inherit',
+      `${pageContainerWidth.dataGraphic}px`,
+    ],
   })}
 `;
 
 const subtitleBottomMargin = css`
   margin-bottom: 24px;
-  margin-left: 24px;
-`;
-
-const tableBottomMargin = css`
-  margin-bottom: 56px;
+  ${mq({
+    marginLeft: ['24px', 'unset'],
+  })}
 `;
 
 const verticalAlign = css`
@@ -131,6 +139,37 @@ function getTableData(rows: Table['children']): Array<TableDataInterface> {
   return rowMap as Array<TableDataInterface>;
 }
 
+function Wrapper({
+  children,
+  isTouchDevice,
+}: {
+  children: React.ReactNode;
+  isTouchDevice: boolean;
+}) {
+  if (isTouchDevice) {
+    return (
+      <div
+        className={css`
+          margin-bottom: 56px;
+        `}
+      >
+        {children}
+      </div>
+    );
+  }
+
+  return (
+    <Card
+      className={css`
+        padding: 24px;
+        margin-bottom: 56px;
+      `}
+    >
+      {children}
+    </Card>
+  );
+}
+
 function PropTable({
   markdownAst,
   component,
@@ -138,6 +177,11 @@ function PropTable({
   markdownAst: ReadmeMarkdown;
   component: string;
 }) {
+  const viewport = useViewportSize();
+
+  const isTouchDevice =
+    viewport !== null ? viewport.width < breakpoints.Tablet : false;
+
   let peerDepIndex: number | undefined;
 
   const headers = markdownAst.children
@@ -211,42 +255,43 @@ function PropTable({
             </Subtitle>
 
             {tableData[index] && (
-              <Table
-                className={tableBottomMargin}
-                key={header}
-                data={tableData[index]}
-                columns={[
-                  <TableHeader dataType="string" label="Prop" key="prop" />,
-                  <TableHeader dataType="string" label="Type" key="type" />,
-                  <TableHeader
-                    dataType="string"
-                    label="Description"
-                    key="description"
-                  />,
-                  <TableHeader
-                    dataType="string"
-                    label="Default"
-                    key="default"
-                  />,
-                ]}
-              >
-                {({ datum }) => (
-                  <Row key={datum.prop.value}>
-                    <Cell className={verticalAlign}>{formatProp(datum)}</Cell>
-                    <Cell className={verticalAlign}>
-                      {typeof datum.type.value === 'string'
-                        ? formatType(datum.type.value, datum.type.url)
-                        : null}
-                    </Cell>
-                    <Cell className={verticalAlign}>
-                      {datum.description.value}
-                    </Cell>
-                    <Cell className={verticalAlign}>
-                      {formatDefault(datum)}
-                    </Cell>
-                  </Row>
-                )}
-              </Table>
+              <Wrapper isTouchDevice={isTouchDevice}>
+                <Table
+                  key={header}
+                  data={tableData[index]}
+                  columns={[
+                    <TableHeader dataType="string" label="Prop" key="prop" />,
+                    <TableHeader dataType="string" label="Type" key="type" />,
+                    <TableHeader
+                      dataType="string"
+                      label="Description"
+                      key="description"
+                    />,
+                    <TableHeader
+                      dataType="string"
+                      label="Default"
+                      key="default"
+                    />,
+                  ]}
+                >
+                  {({ datum }) => (
+                    <Row key={datum.prop.value}>
+                      <Cell className={verticalAlign}>{formatProp(datum)}</Cell>
+                      <Cell className={verticalAlign}>
+                        {typeof datum.type.value === 'string'
+                          ? formatType(datum.type.value, datum.type.url)
+                          : null}
+                      </Cell>
+                      <Cell className={verticalAlign}>
+                        {datum.description.value}
+                      </Cell>
+                      <Cell className={verticalAlign}>
+                        {formatDefault(datum)}
+                      </Cell>
+                    </Row>
+                  )}
+                </Table>
+              </Wrapper>
             )}
           </div>
         );