+++ packages/table/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @leafygreen-ui/table
 
+## 1.3.3
+
+### Patch Changes
+
+- 7ba9f4a8: Increases shadow when Table is scrollable
+
+## 1.3.2
+
+### Patch Changes
+
+- bf58e7d5: Improves default accessibility of Table component by adding the "scope" attribute to column header elements.
+- Updated dependencies [ba56b1cc]
+  - @leafygreen-ui/icon@8.0.0
+
 ## 1.3.1
 
 ### Patch Changes